feat: 八字 golden 扩至 59 例（12节换月/极端时区/日柱循环/大运顺逆/五鼠遁全干）
新增 bazi_golden3.json（32 例，lunar-typescript 生成）+ TestBaziGolden3_AllLayers：
- 12 节换月边界：清明/立夏/芒种/小暑/立秋/白露/寒露/立冬/大雪/小寒（补立春/惊蛰已有）
- 极端时区：UTC+13(新西兰)/-11(萨摩亚)/+5:45(尼泊尔)——验证公历排盘不受时区影响
- 60 甲子日柱连续循环（8 天，day-loop-1..8）
- 大运顺逆 4 组合（阳男/阳女/阴男/阴女）
- 五鼠遁全 10 日干子时（wushu-ri1..5 覆盖甲己/乙庚/丙辛/丁壬/戊癸）
- 立春/惊蛰后 2 小时（验证月柱稳定，避开 lunar 节气近似误差窗口）

顺带修复：
- SolarTermTime 初始猜测改黄经基准（原 35+ti*15 对非立春节不适用）
- AllSolarTerms terms[1]（小寒）改 year 年（原 year-1 错年）
- baziGoldenCase.TZ 改 float64（支持 +5:45 尼泊尔时区）

验证：14/14 全绿，59 个 golden 子测试全 PASS

+++ internal/engine/bazi/bazi_golden_test.go
@@ -16,7 +16,7 @@ import (
 type baziGoldenCase struct {
 	Name   string  `json:"name"`
 	Solar  string  `json:"solar"`
-	TZ     int     `json:"tz"`
+	TZ     float64 `json:"tz"`
 	Lon    float64 `json:"lon"`
 	Gender string  `json:"gender"`
 	Pillars struct {
